+++ src/components/Pages/CreateCompo/CreateCompo.tsx
@@ -1,6 +1,7 @@
-import React from "react";
+import React, {useState} from "react";
 
 import {
+    AppButton,
     ConfigHeader,
     PlayerNumber,
     roles,
@@ -11,6 +12,61 @@ import {
     werewolfState
 } from "@/IndexImporter";
 import "./CreateCompo.scss";
+
+const NextStepButton: React.FC<{"enoughRole": boolean}> = ({enoughRole}) => {
+    const {balanceScore, composition} = useAppSelector(werewolfState.GameData);
+
+    const roleForComedien = roles
+        .filter((role) => role.comedianCanTake)
+        .map((el) => el.variableName);
+    const comedien = composition.find((el) => el.roleName === "comedien");
+    const isComedien = comedien && comedien.quantity > 0;
+    const comedienSpecialRole = composition.filter((el) => {
+        return el.quantity > 0 && roleForComedien.indexOf(el.roleName) >= 0;
+    }).length;
+
+    const voleur = composition.find((el) => el.roleName === "voleur");
+    const isVoleur = voleur && voleur.quantity > 0;
+    const voleurRoles = composition.filter((el) => {
+        return (
+            el.quantity >= 1 &&
+            [
+                "voleur",
+                "deux-soeurs",
+                "trois-freres"
+            ].indexOf(el.roleName) ===
+                -1
+        );
+    }).length;
+    const voleurValid = isVoleur ? voleurRoles >= 2 : true;
+
+    const validToContinue =
+        voleurValid &&
+        enoughRole &&
+        (!isComedien || (isComedien && comedienSpecialRole >= 3));
+
+    return (
+        <>
+            {!isComedien
+                ? (
+                    <AppButton
+                        classname={validToContinue ? "primary" : "disable"}
+                        goal="/game/create/name"
+                        text="Inscrire le nom des joueurs"
+                        type="navLink"
+                    />
+                )
+                : (
+                    <AppButton
+                        classname={validToContinue ? "primary" : "disable"}
+                        goal="/game/create/comedien"
+                        text="Attribuer des rôles au comédien"
+                        type="navLink"
+                    />
+                )}
+        </>
+    );
+};
 
 export const CreateCompo: React.FC = () => {
     // Use the typed version create in hooks.ts
@@ -79,6 +135,7 @@ export const CreateCompo: React.FC = () => {
                 </p>
                 <RolesChoice />
             </section>
+            <NextStepButton enoughRole={roleSelected === roleNeeded} />
         </div>
     );
 };